Add files via upload
The voucher.py file is supposed to be a a python module that guesses or predicts airtime vouchers for any mobile network operator. However it does not have 100% accuracy, I realized I can achieve that by using some AI tools but  I have not implemented those. Advise on how to go about this would be greatly appreciated.